refactor: page.tsx as server component, Timeline owns all scroll logic
